pyproject.toml
[project]
name = "pepeunit-backend"
version = "0.9.0"
description = "Pepeunit - Federated IoT Platform"
authors = [
    {name = "Ivan Serebrennikov", email = "admin@silberworks.com"}
]
license = {text = "AGPL v3 License"}
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "fastapi>=0.95.0",
    "sqlmodel>=0.0.8",
    "uvicorn>=0.21.1",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.27.0",
    "alembic>=1.10.2",
    "psycopg2-binary>=2.9.5",
    "httpx>=0.23.3",
    "uuid>=1.30",
    "python-multipart>=0.0.6",
    "minio>=7.1.15",
    "pyjwt>=2.7.0",
    "fastapi-filter>=0.6.0",
    "filetype>=1.2.0",
    "jinja2>=3.1.2",
    "pyaes>=1.6.1",
    "gitpython>=3.1.40",
    "fastapi-mqtt>=1.2.0",
    "strawberry-graphql>=0.220.0",
    "gunicorn>=21.2.0",
    "aiogram>=3.6.0",
    "pydantic-settings>=2.2.1",
    "toml>=0.10.2",
    "setuptools>=70.0.0",
    "dill>=0.3.8",
    "fastapi-utilities>=0.2.0",
    "pytest>=8.2.2",
    "pytest-asyncio>=0.23.7",
    "paho-mqtt>=2.1.0",
    "prometheus-fastapi-instrumentator>=7.0.2",
    "cryptography>=44.0.0",
    "psutil>=7.0.0",
    "cachetools>=5.5.2",
    "clickhouse-migrations>=0.8.1",
    "clickhouse-driver>=0.2.9",
    "redis>=5.2.1",
    "pyyaml>=6.0.2",
    "pepeunit-client==0.10.6",
    "pathspec>=0.12.1",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.8.0",
    "ruff>=0.13.2",
]
load = [
    "paho-mqtt>=2.1.0",
    "httpx>=0.28.1",
    "tqdm>=4.67.1",
    "psutil>=7.0.0",
    "locust>=2.33.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
explicit = true

[tool.uv.sources]
pepeunit-client = { index = "testpypi" }

[tool.ruff]
line-length = 79

target-version = "py313"

exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "dist",
    "build",
    "alembic/versions",
    "tests",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear (включает проверку deprecated функций)
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "Q",   # flake8-quotes
    "C90", # mccabe
    "B008", # flake8-bugbear (проверка default arguments)
    "B028", # flake8-bugbear (проверка deprecated функций)
]

ignore = [
    "E501",  # line too long
    "B008",  # Do not perform function calls in argument defaults
    "B006",  # Do not use mutable data structures for argument defaults
    "C901",  # too complex (temporarily disabled)
    "SIM115" # with for files (for lock work disable)
]

[tool.ruff.lint.isort]
# Настройки для сортировки импортов
known-first-party = ["app"]
force-single-line = false
